Restructure interfaces to have one for environment (env::System) and one for translation (gen::Translator)
Rename original Translator to Block and add guest page-size as configurable parameter to env::System

+++ generate/context/context-writer.cpp
@@ -31,7 +31,7 @@ void gen::detail::ContextWriter::fCheckRange(uint32_t offset, gen::MemoryType ty
 		break;
 	}
 
-	uint32_t size = env::detail::ContextAccess::ContextSize();
+	uint32_t size = uint32_t(env::detail::ContextAccess::ContextSize());
 	if (offset > size)
 		host::Fatal(u8"Cannot read [", offset, u8"] bytes from context of size [", size, u8']');
 }
